Move dmi::render to dmi/render.rs, remove AsRef, use a tuple key for state_names
